tools/sfmc_scanner/requirements.txt
requests
sqlparse
python-dotenv
